Employer post and get implementation

+++ src/main/java/com/cmpe275/Lab2/repository/EmployeeRepository.java
@@ -2,6 +2,6 @@ package com.cmpe275.Lab2.repository;
 import com.cmpe275.Lab2.models.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 
 }
